Refactor common code into an AbstractApplicationManager
